Working graphql organisaties

+++ backend/hhb_backend/graphql/mutations/organisaties/delete_organisatie.py
@@ -6,7 +6,7 @@ import requests
 from graphql import GraphQLError
 from hhb_backend.graphql import settings
 
-class DeleteGebruiker(graphene.Mutation):
+class DeleteOrganisatie(graphene.Mutation):
     class Arguments:
         # gebruiker arguments
         id = graphene.Int(required=True)
@@ -15,11 +15,23 @@ class DeleteGebruiker(graphene.Mutation):
 
     def mutate(root, info, id):
         """ Delete current gebruiker """
-
-        delete_response = requests.delete(
-            os.path.join(settings.HHB_SERVICES_URL, f"gebruikers/{id}")
+        kvk_nummer_response = requests.get(
+            f"{settings.HHB_SERVICES_URL}/organisaties/{id}"
         )
-        if delete_response.status_code != 204:
-            raise GraphQLError(f"Upstream API responded: {delete_response.json()}")
-        return DeleteGebruiker(ok=True)
+        if kvk_nummer_response.status_code != 200:
+            raise GraphQLError(f"Upstream API responded: {kvk_nummer_response.json()}")
+        kvk_nummer = kvk_nummer_response.json()["data"]["kvk_nummer"]
+
+        delete_response_hhb = requests.delete(
+            os.path.join(settings.HHB_SERVICES_URL, f"organisaties/{id}")
+        )
+        if delete_response_hhb.status_code != 202:
+            raise GraphQLError(f"Upstream API responded: {delete_response_hhb.json()}")
+
+        delete_response_org = requests.delete(
+            os.path.join(settings.ORGANISATIE_SERVICES_URL, f"organisaties/{kvk_nummer}")
+        )
+        if delete_response_org.status_code not in [202, 404]:
+            raise GraphQLError(f"Upstream API responded: {delete_response_hhb.json()}")
+        return DeleteOrganisatie(ok=True)
